+++ frontend/src/components/Login/SignUpButton.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Button } from './style';
+
+const SignUpButton: React.FC = () => {
+	return (
+		<Button>
+			<span>Signup</span>
+		</Button>
+	);
+};
+
+export default SignUpButton;
