+++ frontend/src/styles/global.ts
@@ -3,6 +3,9 @@ import reset from 'styled-reset';
 
 const GlobalStyle = createGlobalStyle`
   ${reset}
+  a {
+    text-decoration: none;
+  }
 `;
 
 export default GlobalStyle;
